carbon: move state partition to 3rd sector and reduce size
mcuboot is now larger than 32k which means that any time we write the state
information, we're overwriting the end of the bootloader sector.

To avoid this, let's move the state information to the 3rd sector in STM32F4
flash layout which has 0xc000 offset and a size of 16k.

Signed-off-by: Michael Scott <michael.scott@linaro.org>