bluemix: decouple DM from core bluemix
We want to use Bluemix publishing data, and only optionally for device
management. To start allowing for this, always compile in the bluemix
core, but only advertise device management support to the serve if
configured to do so.

This doesn't finish the job, however. To do that, Bluemix and hawkBit
will need separate net_context objects, then main.c will need to
handle simultaneous use of both.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>