hawkbit: don't wait forever for data from hawkbit
In testing, it seems the tinyproxy process running on the
IoT gateways can get into an odd state where connections are
accepted, but a full HTTP session is never completed for
some reason.   When this happens, all of the threads which are
using the managed tcp interface log will all be hung waiting
for this hawkbit query to finish.

Best to timeout after a reasonable period and free up the TCP
interface lock so that other threads can continue.

TODO: Needs more debugging to see how tinyproxy ends up in
this state.

Signed-off-by: Michael Scott <michael.scott@linaro.org>