dm-lwm2m: Add support for LWM2M credentials.
We want to support secure FOTA via LWM2M. To do that, we'll need to
access persistent data which contain LWM2M-specific configuration data,
such as the Device ID and Device Token (DTLS).

Use defines generated by a DTS partition to access this information,
which must be provisioned per-device. As long as the partitions are
configured properly not to overlap with mcuboot or its image and
scratch areas, and this flash partition is not erased by this
application, it will persist.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>
Signed-off-by: Ricardo Salveti <ricardo.salveti@linaro.org>