dm-lwm2m: select NET_APP_NEED_IPV6 for Nitrogen/Carbon
This setting is used during the net_app framework network init.
Currently, we immediately timeout and let the app make sure that
networking is up correctly, due to the way bluetooth networking
is being initialized.  This means that this setting isn't
actually used.  However, in the future we may revert to letting
the net_app init work normally for bluetooth devices and this
setting will be required.

Let's fix this now rather than hunt for it later.

Signed-off-by: Michael Scott <michael.scott@linaro.org>